feat: add ability to set ships' positions
ships have coordinates, that are an array of 2 elements - first one being the x index (col idx), second one being y index (row idx). a ship's coordinates can be set using the .setCoordinates() method in ships. this is called in the gameboard factory in the .placeShip() method. we ensure that ships do not go out of bounds, and also players cannot place a ship where there is already a ship there. by creating a hashmap of ship types valid in the 2002 version of Hasbro's battleship, we ensure we don't have non-standard ships being created.