refactor: coordinatesWithShip is now a hashmap that contains key being the coordinate (so it can be looked up quickly to check if a ship has been hit), and value being the ship type. remove unused getter in ship factory
i did not make the entire ship object the value, as, once placed, we will then need to revisit the  hashmap to confirm if it has been sunk. it's better to do this as it is leaner, and avoids issues with the object being properly updated. but i might be wrong. let's see.