feat: user can click on the computer/opposition's grid squares to attack it. DOM shows whether square was a hit or miss immediately, but also updates the objects, all the way down into the individual square's properties so that it can be re-rendered.
important to note here that, on click, the DOM is not directly manipulated by screen controller. instead, we update the underlying models (squares, gameboards, players etc), and then this gets re-rendered by the screenController's .updateGrid method. this maintains a clear separation of concerns between the logic (whether a square has a ship in it AND got hit or not) and what is displayed (which pulls from that underlying, rendered board)